requirements.txt
Django==3.2.3
djangorestframework==3.12.4
requests==2.32.3
tinydb==4.7.0